# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: keen-pbr API
  description: REST API for the keen-pbr policy-based routing daemon.
  version: "3.0.0"
servers:
  - url: http://127.0.0.1:12121
    description: Example local API address. Runtime default listen is `0.0.0.0:12121`.
paths:
...
  /api/transports:
    get:
      summary: Transport manager runtime state
      description: >
        Proxies the loopback-only transport manager API and returns the
        current state of native and managed tunnel interfaces.
      operationId: getTransports
      responses:
        "200":
          description: Transport runtime states
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransportStatus"
        "503":
          description: Transport manager is not configured or unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Start or stop a transport
      operationId: postTransportAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportActionRequest"
      responses:
        "200":
          description: Action accepted by the transport manager
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportActionResponse"
        "400":
          description: Invalid tag or action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Transport manager is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    TransportStatus:
      type: object
      required: [tag, type, interface, state, updated_at, desired_up]
      properties:
        tag:
          type: string
        type:
          type: string
        interface:
          type: string
        state:
          type: string
          enum: [down, starting, up, degraded]
        pid:
          type: integer
        error:
          type: string
        updated_at:
          type: string
          format: date-time
        desired_up:
          type: boolean
          description: Whether the supervisor should keep this managed transport up
        retry_count:
          type: integer
          minimum: 0
        next_retry_at:
          type: string
          format: date-time
    TransportActionRequest:
      type: object
      required: [tag, action]
      properties:
        tag:
          type: string
          pattern: "^[a-z][a-z0-9_]{0,23}$"
        action:
          type: string
          enum: [up, down]
    TransportActionResponse:
      type: object
      required: [status, at]
      properties:
        status:
          type: string
          enum: [accepted]
        at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error message.
          example: "Cannot open config file"
        validation_errors:
          type: array
          description: Optional list of validation failures returned by `POST /api/config`.
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      required: [message]
      properties:
        path:
          type: string
          description: Dot-style config path associated with the validation failure.
          example: "fwmark.mask"
        message:
          type: string
          description: Human-readable validation message.
          example: "fwmark.mask must have one or more consecutive hex nibbles set to F (e.g. 0x000F0000, 0x00FF0000, 0x0FFF0000), got 0x0f0f0000"

    # -------------------------------------------------------------------------
    # /api/health/routing
    # -------------------------------------------------------------------------
...